Add Storage class
Add initial format for Storage class, with a constructor implementation that creates a new local json file if none exists already.

Two functions will be available for use;
* Get
* Post

Which will work similar to most APIs where we are able to get and post into the json storage object.

Related to issue #17