Feat - Reads the output of the bash commands
The commit implements features that reads the output, errors and exit values of bash commands.
The output of each command are stored in ArrayLists which in turn are stored in a bigger ArrayList.
Each exit value is stored as the last element in each command-list and can easily be taken out and stored separately.
The large ArrayList is returned to the main class and a function prints the output for temporary visual feedback.