Integrate Build class into storage
Storage and API skeleton code are now working together. API is now able to return a list of Build objects that should be fetched from the frontend. This is done by performing a get request to <your local address>/ci/get and you will get a reply in the following format:

* jobID
  * status
  * commitSha
  * url
  * log

Add getAll function in Storage class so that all builds can be fetched at the same time for performance and ease of use reasons.

Related to issue #26