Implement local storage (#26)
* Add Storage class

Add initial format for Storage class, with a constructor implementation that creates a new local json file if none exists already.

Two functions will be available for use;
* Get
* Post

Which will work similar to most APIs where we are able to get and post into the json storage object.

Related to issue #17

* Add post function

Post function now works as intended and adds the storage object (with a maybe to change format) to the database.

Related to issue #17 and #25

* Add Get function

Get function should now work as intended and fetches a given object from the database depending on its id. If key was not found in object, then return an empty object.

Related to issue #17 and #25

* Add gradle dependencies

Add some new gradle dependencies that were used for Storage

* Integrate Build class into storage

Storage and API skeleton code are now working together. API is now able to return a list of Build objects that should be fetched from the frontend. This is done by performing a get request to <your local address>/ci/get and you will get a reply in the following format:

* jobID
  * status
  * commitSha
  * url
  * log

Add getAll function in Storage class so that all builds can be fetched at the same time for performance and ease of use reasons.

Related to issue #26

* Fix naming

Naming was broken after merge, should be fixed now.

* Refactor Resource.java

Move the chunk of code that converts a json object to a Build object into a seperate function