Make frontend dynamic
Frontend now fetches from the database and creates a list item for each one. It does not contain the exact information we want to show yet however, so it is not finished.

Added dotenv to store ngrok and proxy urls.

In order to run, add your ngrok url (not localhost) to the .env file and then start the frontend by running 'npm start'. Other than this, the ngrok server and backend server also have to be running.

Updated .gitignore to hide .env file which means you have to create your own in order to start the server. Ask me for more information about how to set it up.

This is related to issue #24 and #40